+++ .github/workflows/update_sdk_methods.py
@@ -1,15 +1,16 @@
 from bs4 import BeautifulSoup
 from urllib.request import urlopen
 from pathlib import Path
-from markdownify import markdownify as md
 import sys
 import os
 import subprocess
-import urllib.parse
-import urllib.error
 import re as regex
 import argparse
 
+from parse_python import PythonParser
+from parse_go import GoParser
+from parse_flutter import FlutterParser
+from parser_utils import make_soup
 
 ## The full list of SDK languages we scrape. You can use the sdk_languages
 ## positional parameter to refine this at runtime if desired:
@@ -323,118 +324,6 @@ proto_map = {
     }
 }
 
-## Language-specific resource name overrides:
-##   "proto_resource_name" : "language-specific_resource_name"
-##   "as-it-appears-in-type-array": "as-it-is-used-per-sdk"
-## Note: Always remap generic component and service, for all languages,
-##       as this must be unique for this script, but is non-unique across sdks.
-go_resource_overrides = {
-    "generic_component": "generic",
-    "input_controller": "input",
-    "movement_sensor": "movementsensor",
-    "power_sensor": "powersensor",
-    "generic_service": "generic",
-    "base_remote_control": "baseremotecontrol",
-    "data_manager": "datamanager"
-}
-
-## Ignore these specific APIs if they error, are deprecated, etc:
-## {resource}.{methodname} to exclude a specific method, or
-## interface.{interfacename} to exclude an entire Go interface:
-go_ignore_apis = [
-    'interface.NavStore', # motion service interface
-    'interface.LocalRobot', # robot interface
-    'interface.RemoteRobot', # robot interface
-    'robot.RemoteByName', # robot method
-    'robot.ResourceByName', # robot method
-    'robot.RemoteNames', # robot method
-    #'robot.ResourceNames', # robot method
-    'robot.ResourceRPCAPIs', # robot method
-    'robot.ProcessManager', # robot method
-    'robot.OperationManager', # robot method
-    'robot.SessionManager', # robot method
-    'robot.PackageManager', # robot method
-    'robot.Logger' # robot method
-]
-
-## Use these URLs for data types (for params, returns, and errors raised) that are
-## built-in to the language or provided by a non-Viam third-party package:
-## TODO: Not currently using these in parse(), but could do a simple replace()
-##       or could handle in markdownification instead. TBD. Same with other SDK lang link arrays:
-go_datatype_links = {
-    "context": "https://pkg.go.dev/context",
-    "map": "https://go.dev/blog/maps",
-    "bool": "https://pkg.go.dev/builtin#bool",
-    "int": "https://pkg.go.dev/builtin#int",
-    "float64": "https://pkg.go.dev/builtin#float64",
-    "image": "https://pkg.go.dev/image#Image",
-    "r3.vector": "https://pkg.go.dev/github.com/golang/geo/r3#Vector",
-    "string": "https://pkg.go.dev/builtin#string",
-    "*geo.Point": "https://pkg.go.dev/github.com/kellydunn/golang-geo#Point",
-    "primitive.ObjectID": "https://pkg.go.dev/go.mongodb.org/mongo-driver/bson/primitive#ObjectID",
-    "error": "https://pkg.go.dev/builtin#error"
-}
-
-## Language-specific resource name overrides:
-python_resource_overrides = {
-    "generic_component": "generic",
-    "input_controller": "input",
-    "generic_service": "generic",
-    "data": "data_client",
-    "app": "app_client",
-    "billing": "billing_client",
-    "data": "data_client",
-    ## Python bundles Dataset and Datasync protos in with Data,
-    ## while Flutter does not. HACK:
-    "dataset": "data_client",
-    "data_sync": "data_client",
-    "mltraining": "ml_training_client"
-}
-
-## Ignore these specific APIs if they error, are deprecated, etc:
-python_ignore_apis = [
-    'viam.app.app_client.AppClient.get_rover_rental_robots', # internal use
-    'viam.app.app_client.AppClient.get_rover_rental_parts', # internal use
-    'viam.app.data_client.DataClient.create_filter', # deprecated
-    'viam.app.data_client.DataClient.delete_tabular_data_by_filter', # deprecated
-    'viam.components.input.client.ControllerClient.reset_channel', # GUESS ?
-    'viam.robot.client.RobotClient.transform_point_cloud', # unimplemented
-    'viam.robot.client.RobotClient.get_component', # GUESS ?
-    'viam.robot.client.RobotClient.get_service', # GUESS ?
-    'viam.components.board.client.BoardClient.write_analog', # Currently borked: https://python.viam.dev/autoapi/viam/components/board/client/index.html#viam.components.board.client.BoardClient.write_analog
-    'viam.components.board.client.StreamWithIterator.next', # No content upstream
-    'viam.robot.client.ViamChannel.close', # channel-specific close
-    'viam.robot.client.SessionsClient.reset'  # session-specific reset
-]
-
-## Use these URLs for data types that are not otherwise captured by parse(), such as:
-## - Well-known built-in data types that are not scrapeable (like 'int')
-## - Viam-specific data types, even if scrapeable, that are part of a multiple-data-type return
-##   (like list_organization_members : Tuple[List[viam.proto.app.OrganizationMember], List[viam.proto.app.OrganizationInvite]]
-## Data type links defined here will be used instead of scraped links if both exist:
-python_datatype_links = {
-    ## Built-in data types:
-    "str": "https://docs.python.org/3/library/stdtypes.html#text-sequence-type-str",
-    "int": "https://docs.python.org/3/library/stdtypes.html#numeric-types-int-float-complex",
-    "float": "https://docs.python.org/3/library/stdtypes.html#numeric-types-int-float-complex",
-    "bytes": "https://docs.python.org/3/library/stdtypes.html#bytes-objects",
-    "bool": "https://docs.python.org/3/library/stdtypes.html#boolean-type-bool",
-    "datetime.datetime": "https://docs.python.org/3/library/datetime.html",
-    "datetime.timedelta": "https://docs.python.org/3/library/datetime.html#timedelta-objects",
-    ## Third-party data types:
-    "numpy.typing.NDArray": "https://numpy.org/doc/stable/reference/typing.html#numpy.typing.NDArray",
-    ## Viam-specific data types:
-    "viam.proto.app.OrganizationMember": "https://python.viam.dev/autoapi/viam/proto/app/index.html#viam.proto.app.OrganizationMember",
-    "viam.proto.app.OrganizationInvite": "https://python.viam.dev/autoapi/viam/proto/app/index.html#viam.proto.app.OrganizationInvite",
-    "viam.components.arm.KinematicsFileFormat.ValueType": "https://python.viam.dev/autoapi/viam/components/arm/index.html#viam.components.arm.KinematicsFileFormat",
-    "viam.media.video.NamedImage": "https://python.viam.dev/autoapi/viam/media/video/index.html#viam.media.video.NamedImage",
-    "viam.proto.common.ResponseMetadata": "https://python.viam.dev/autoapi/viam/gen/common/v1/common_pb2/index.html#viam.gen.common.v1.common_pb2.ResponseMetadata",
-    "viam.proto.component.encoder.PositionType.ValueType": "https://python.viam.dev/autoapi/viam/gen/component/encoder/v1/encoder_pb2/index.html#viam.gen.component.encoder.v1.encoder_pb2.PositionType",
-    "typing_extensions.Self": "https://python.viam.dev/autoapi/viam/robot/client/index.html#viam.robot.client.RobotClient",
-    "viam.components.movement_sensor.movement_sensor.MovementSensor.Accuracy": "https://python.viam.dev/autoapi/viam/components/movement_sensor/movement_sensor/index.html#viam.components.movement_sensor.movement_sensor.MovementSensor.Accuracy",
-    "viam.services.vision.vision.Vision.Properties": "https://python.viam.dev/autoapi/viam/components/audio_input/audio_input/index.html#viam.components.audio_input.audio_input.AudioInput.Properties"
-}
-
 ## Inject these URLs, relative to 'docs', into param/return/raises descriptions that contain exact matching key text.
 ## write_markdown() uses override_description_links via link_description() when it goes to write out descriptions.
 ## Currently only used for method descriptions, but see commented-out code for usage as optional consideration.
@@ -455,26 +344,6 @@ override_description_links = {
     "API key": "/fleet/cli/#authenticate",
     "board model": "/dev/reference/apis/components/board/"
 }
-
-## Language-specific resource name overrides:
-flutter_resource_overrides = {
-    "generic_component": "Generic",
-    "movement_sensor": "MovementSensor",
-    "power_sensor": "PowerSensor",
-    "vision": "VisionClient",
-    "robot": "RobotClient",
-    "data": "DataClient",
-    "dataset": "DataClient",
-    "data_sync": "DataClient"
-}
-
-## Ignore these specific APIs if they error, are deprecated, etc:
-flutter_ignore_apis = [
-    'getRoverRentalRobots' # internal use
-]
-
-## Use these URLs for data types that are built-in to the language:
-flutter_datatype_links = {}
 
 ## Map sdk language to specific code fence formatting syntax for that language:
 code_fence_fmt = {
@@ -572,66 +441,6 @@ def get_proto_apis():
 
     return proto_map
 
-## Fetch URL content via BS4, used in parse():
-def make_soup(url):
-   try:
-       page = urlopen(url)
-       html = page.read().decode("utf-8")
-       return BeautifulSoup(html, "html.parser")
-   except urllib.error.HTTPError as err:
-       print(f'An HTTPError was thrown: {err.code} {err.reason} for URL: {url}')
-
-def shorten_data_type(t):
-    if '.' in t:
-        return '.'.join(t.split('.')[-2:])
-    else:
-        return t
-
-## Link any matching data types to their reference links, based on {sdk}_datatype_links[] array,
-## used in parse() for both param and return data types. Handles data types syntax that includes
-## multiple data types (and therefore requires multiple data type links), such as
-## ListOrganizationMembers: Tuple[List[viam.proto.app.OrganizationMember], List[viam.proto.app.OrganizationInvite]
-## DESIGN DECISION: Ignore well-known, usually leading (containing) data types like List, Tuple, Dict.
-## NOTE: Only used in PySDK parsing, for now (but should work for all with minor tweak to support per-language links array):
-def link_data_types(sdk, data_type_string):
-
-    linked_data_type_string = ""
-
-    ## If the passed data_type_string matches exactly to a data type defined in python_datatype_links, use that:
-    if data_type_string in python_datatype_links.keys():
-        shorter_data_type = shorten_data_type(data_type_string)
-        linked_data_type_string = '[' + shorter_data_type + '](' + python_datatype_links[data_type_string] + ')'
-    else:
-
-        ## Assemble all encountered data types that match to python_datatype_links keys into array.
-        ## This match is a little too greedy, and will match, say, 'int' to 'JointPositions'. To counter
-        ## this, we additionally check for leading and trailing alphanumeric characters further in:
-        matching_data_types = list(key for key in python_datatype_links if key in data_type_string)
-
-        if len(matching_data_types) > 0:
-
-            ## Ugly hack to allow us to append within the for loop below, sorry:
-            linked_data_type_string = data_type_string
-
-            for data_type_found in matching_data_types:
-
-                ## Discard string matches that are substrings of other data type strings:
-                if not regex.search(r'[A-Za-z0-9]' + data_type_found, data_type_string) and not regex.search(data_type_found + r'[A-Za-z0-9]', data_type_string):
-
-                    shorter_data_type = shorten_data_type(data_type_found)
-                    data_type_linked = '[' + shorter_data_type + '](' + python_datatype_links[data_type_found] + ')'
-                    linked_data_type_string = regex.sub(data_type_found, data_type_linked, linked_data_type_string)
-                else:
-                    ## If we get here, this data_type is actually a substring of another data type. Take no action:
-                    pass
-
-    ## If we didn't find any matching links, return an empty string so we can know to look elsewhere,
-    ## otherwise return linked data type string:
-    if linked_data_type_string == data_type_string:
-        return ""
-    else:
-        return linked_data_type_string
-
 
 ## Link matching text, used in write_markdown():
 ## NOTE: Currently does not support formatting for link titles
@@ -672,893 +481,16 @@ def parse(type, names):
         ## Build empty dict to house methods, and update scrape_url
         ## with staging link if we detected one earlier:
         if sdk == "go":
-            go_methods = {}
-            go_methods[type] = {}
-            if go_staging_url != '':
-                scrape_url = go_staging_url
+            go_parser = GoParser(proto_map_file, go_staging_url)
+            all_methods["go"] = go_parser.parse(type, names)
         elif sdk == "python":
-            python_methods = {}
-            python_methods[type] = {}
-            if python_staging_url != '':
-                scrape_url = python_staging_url
+            python_parser = PythonParser(proto_map_file, python_staging_url)
+            all_methods["python"] = python_parser.parse(type, names)
         elif sdk == "flutter":
-            flutter_methods = {}
-            flutter_methods[type] = {}
-            flutter_methods[type][resource] = {}
-            if flutter_staging_url != '':
-                scrape_url = flutter_staging_url
+            flutter_parser = FlutterParser(proto_map_file, flutter_staging_url)
+            all_methods["flutter"] = flutter_parser.parse(type, names, args)
         else:
             print("unsupported language!")
-
-        ## Iterate through each resource (like 'arm') in type (like 'components') array:
-        for resource in names:
-
-            ## Determine URL form for Go depending on type (like 'component'):
-            if sdk == "go":
-                if type in ("component", "service") and resource in go_resource_overrides:
-                    url = f"{scrape_url}/go.viam.com/rdk/{type}s/{go_resource_overrides[resource]}"
-                elif type in ("component", "service"):
-                    url = f"{scrape_url}/go.viam.com/rdk/{type}s/{resource}"
-                elif type == "robot" and resource in go_resource_overrides:
-                    url = f"{scrape_url}/go.viam.com/rdk/{type}/{go_resource_overrides[resource]}"
-                elif type == "robot":
-                    url = f"{scrape_url}/go.viam.com/rdk/{type}"
-                elif type == "app":
-                    pass
-                go_methods[type][resource] = {}
-
-            ## Determine URL form for Python depending on type (like 'component'):
-            elif sdk == "python":
-                if type in ("component", "service") and resource in python_resource_overrides:
-                    url = f"{scrape_url}/autoapi/viam/{type}s/{python_resource_overrides[resource]}/client/index.html"
-                elif type in ("component", "service"):
-                    url = f"{scrape_url}/autoapi/viam/{type}s/{resource}/client/index.html"
-                elif type == "app" and resource in python_resource_overrides:
-                    url = f"{scrape_url}/autoapi/viam/{type}/{python_resource_overrides[resource]}/index.html"
-                elif type == "app":
-                    url = f"{scrape_url}/autoapi/viam/{type}/{resource}/index.html"
-                else: # robot
-                    url = f"{scrape_url}/autoapi/viam/{type}/client/index.html"
-                python_methods[type][resource] = {}
-
-            ## Determine URL form for Flutter depending on type (like 'component').
-            ## TEMP: Manually exclude Base Remote Control Service (Go only):
-            ## TODO: Handle resources with 0 implemented methods for this SDK better.
-            elif sdk == "flutter":
-                # Initialize Flutter methods dictionary if it doesn't exist
-                if 'flutter' not in all_methods:
-                    all_methods['flutter'] = {}
-                if type not in all_methods['flutter']:
-                    all_methods['flutter'][type] = {}
-
-                # Initialize empty dictionary for this resource
-                all_methods['flutter'][type][resource] = {}
-
-                # Skip resources not supported in Flutter
-                unsupported_resources = [
-                    "base_remote_control", "button", "encoder", "input_controller",
-                    "data_manager", "generic_service", "mlmodel", "motion",
-                    "navigation", "slam", "switch"
-                ]
-                if resource in unsupported_resources:
-                    if args.verbose:
-                        print(f'DEBUG: Skipping unsupported Flutter resource: {resource}')
-                    continue
-                elif resource in flutter_resource_overrides:
-                    url = f"{scrape_url}/viam_sdk/{flutter_resource_overrides[resource]}-class.html"
-                    if args.verbose:
-                        print(f'DEBUG: Parsing Flutter URL: {url}')
-                else:
-                    url = f"{scrape_url}/viam_sdk/{resource.capitalize()}-class.html"
-            ## If an invalid language was provided:
-            else:
-                pass
-
-            ## Scrape each parent method tag and all contained child tags for Go by resource:
-            ## Skip Go: App (Go has no App client) and the generic component and service, which
-            ## require explicit in-script workaround (DoCommand neither inherited (from resource.Resource)
-            ## nor explicitly defined in-interface (not in Go docs, only in un-doc'd code):
-            if sdk == "go" and type != "app" and resource != "generic_component" and resource != "generic_service":
-
-                soup = make_soup(url)
-
-                ## Get a raw dump of all go methods by interface for each resource:
-                go_methods_raw = soup.find_all(
-                    lambda tag: tag.name == 'div'
-                    and tag.get('class') == ['Documentation-declaration']
-                    and tag.pre.text.startswith('type')
-                    and "interface {" in tag.pre.text)
-
-                # some resources have more than one interface:
-                for resource_interface in go_methods_raw:
-
-                    ## Determine the interface name, which we need for the method_link:
-                    interface_name = resource_interface.find('pre').text.splitlines()[0].removeprefix('type ').removesuffix(' interface {')
-
-                    ## Exclude unwanted Go interfaces:
-                    check_interface_name = 'interface.' + interface_name
-                    if not check_interface_name in go_ignore_apis:
-
-                        ## Loop through each method found for this interface:
-                        for tag in resource_interface.find_all('span', attrs={"data-kind" : "method"}):
-
-                            ## Create new empty dictionary for this specific method, to be appended to ongoing go_methods dictionary,
-                            ## in form: go_methods[type][resource][method_name] = this_method_dict
-                            this_method_dict = {}
-
-                            tag_id = tag.get('id')
-                            method_name = tag.get('id').split('.')[1]
-
-                            ## Exclude unwanted Go methods:
-                            check_method_name = resource + '.' + method_name
-                            if not check_method_name in go_ignore_apis:
-
-                                ## Look up method_name in proto_map file, and return matching proto:
-                                with open(proto_map_file, 'r') as f:
-                                    for row in f:
-                                        if not row.startswith('#') \
-                                        and row.startswith(resource + ',') \
-                                        and row.split(',')[4] == method_name:
-                                            this_method_dict["proto"] = row.split(',')[1]
-
-                                ## Extract the raw text from resource_interface matching method_name.
-                                ## Split by method span, throwing out remainder of span tag, catching cases where
-                                ## id is first attr or data-kind is first attr, and slicing to omit the first match,
-                                ## which is the opening of the method span tag, not needed:
-                                this_method_raw1 = regex.split(r'id="' + tag_id + '"', str(resource_interface))[1].removeprefix('>').removeprefix(' data-kind="method">').lstrip()
-
-                                ## Then, omit all text that begins a new method span, and additionally remove trailing
-                                ## element closers for earlier tags we spliced into (pre and span):
-                                this_method_raw2 = regex.split(r'<span .*data-kind="method".*>', this_method_raw1)[0].removesuffix('}</pre>\n</div>').removesuffix('</span>').rstrip()
-
-                                method_description = ""
-
-                                ## Get method description, if any comment spans are found:
-                                if tag.find('span', class_='comment'):
-
-                                    ## Iterate through all comment spans, splitting by opening comment tag, and
-                                    ## omitting the first string, which is either the opening comment tag itself,
-                                    ## or the usage of this method, if the comment is appended to the end of usage line:
-                                    for comment in regex.split(r'<span class="comment">', this_method_raw2)[1:]:
-
-                                        comment_raw = regex.split(r'</span>.*', comment.removeprefix('//'))[0].lstrip()
-                                        method_description = method_description + comment_raw
-
-                                ## Write comment field as appended comments if found, or empty string if none.
-                                this_method_dict["description"] = method_description
-
-                                ## Get full method usage string, by omitting all comment spans:
-                                method_usage_raw = regex.sub(r'<span class="comment">.*</span>', '', this_method_raw2)
-                                method_usage_raw2 = regex.sub(r'</span>', '', method_usage_raw).replace("\t", "  ").lstrip().rstrip()
-                                ## Some Go params use versioned links, some omit the version (to use latest).
-                                ## Standardize on using latest for all cases. This handles parameters and returns:
-                                this_method_dict["usage"] = regex.sub(r'/rdk@v[0-9\.]*/', '/rdk/', method_usage_raw2, flags=regex.DOTALL)
-
-                                ## Not possible to link to the specific functions, so we link to the parent resource instead.
-                                ## If we are scraping from a local staging instance, replace host and port with upstream link target URL:
-                                if go_staging_url != '':
-                                    this_method_dict["method_link"] = str(url + '#' + interface_name).replace(go_staging_url, 'https://pkg.go.dev')
-                                else:
-                                    this_method_dict["method_link"] = url + '#' + interface_name
-
-                                ## Check for code sample for this method.
-                                go_code_samples_raw = soup.find_all(
-                                    lambda code_sample_tag: code_sample_tag.name == 'p'
-                                    and code_sample_tag.text.startswith(method_name)
-                                    and code_sample_tag.text.endswith(method_name + ' example:\n'))
-
-                                ## Determine if a code sample is provided for this method:
-                                if len(go_code_samples_raw) == 1:
-
-                                    ## Fetch code sample raw text, preserving newlines but stripping all formatting.
-                                    ## This string should be suitable for feeding into any python formatter to get proper form:
-                                    this_method_dict["code_sample"] = go_code_samples_raw[0].find_next('pre').text.replace("\t", "  ")
-
-                                elif len(go_code_samples_raw) > 1:
-
-                                    ## In case we want to support multiple code samples per method down the line,
-                                    ## this is where to process (and: update write_markdown() accordingly to enable looping
-                                    ## through possible code sample data objects). For now we just continue to fetch just the
-                                    ## first-discovered (i.e., at index [0]):
-                                    this_method_dict["code_sample"] = go_code_samples_raw[0].find_next('pre').text.replace("\t", "  ")
-
-                                ## We have finished collecting all data for this method. Write the this_method_dict dictionary
-                                ## in its entirety to the go_methods dictionary by type (like 'component'), by resource (like 'arm'),
-                                ## using the method_name as key:
-                                go_methods[type][resource][method_name] = this_method_dict
-
-                        ## If this Go interface inherits from another interface, also fetch data for those inherited methods:
-                        if '\tresource.' in resource_interface.text:
-
-                            resource_url = f"{scrape_url}/go.viam.com/rdk/resource"
-                            resource_soup = make_soup(resource_url)
-
-                            ## If the resource being considered inherits from resource.Resource (currently all components and services do,
-                            ## and no app or robot interfaces do), then add the three inherited methods manually: Reconfigure(), DoCommand(), Close()
-                            if '\tresource.Resource' in resource_interface.text:
-                                go_methods[type][resource]['Reconfigure'] = {'proto': 'Reconfigure', \
-                                    'description': 'Reconfigure must reconfigure the resource atomically and in place. If this cannot be guaranteed, then usage of AlwaysRebuild or TriviallyReconfigurable is permissible.', \
-                                    'usage': 'Reconfigure(ctx <a href="/context">context</a>.<a href="/context#Context">Context</a>, deps <a href="#Dependencies">Dependencies</a>, conf <a href="#Config">Config</a>) <a href="/builtin#error">error</a>', \
-                                    'method_link': 'https://pkg.go.dev/go.viam.com/rdk/resource#Resource'}
-                                code_sample = resource_soup.find_all(lambda code_sample_tag: code_sample_tag.name == 'p' and "Reconfigure example:" in code_sample_tag.text)
-                                if code_sample:
-                                    go_methods[type][resource]['Reconfigure']['code_sample'] = code_sample[0].find_next('pre').text.replace("\t", "  ")
-                                go_methods[type][resource]['DoCommand'] = {'proto': 'DoCommand', \
-                                    'description': 'DoCommand sends/receives arbitrary data.', \
-                                    'usage': 'DoCommand(ctx <a href="/context">context</a>.<a href="/context#Context">Context</a>, cmd map[<a href="/builtin#string">string</a>]interface{}) (map[<a href="/builtin#string">string</a>]interface{}, <a href="/builtin#error">error</a>)', \
-                                    'method_link': 'https://pkg.go.dev/go.viam.com/rdk/resource#Resource'}
-                                code_sample = resource_soup.find_all(lambda code_sample_tag: code_sample_tag.name == 'p' and "DoCommand example:" in code_sample_tag.text)
-                                if code_sample:
-                                    if type == "component":
-                                        go_methods[type][resource]['DoCommand']['code_sample'] = 'my' + resource.title().replace("_", "")+ ', err := ' + resource + '.FromRobot(machine, "my_' + resource + '")\n\ncommand := map[string]interface{}{"cmd": "test", "data1": 500}\nresult, err := my' + resource.title().replace("_", "") + '.DoCommand(context.Background(), command)\n'
-                                        if resource == "generic_component":
-                                            go_methods[type][resource]['DoCommand']['code_sample'] = 'myGenericComponent, err := generic.FromRobot(machine, "my_generic_component")\n\ncommand := map[string]interface{}{"cmd": "test", "data1": 500}\nresult, err := myGenericComponent.DoCommand(context.Background(), command)\n'
-                                    else:
-                                        go_methods[type][resource]['DoCommand']['code_sample'] = 'my' + resource.title().replace("_", "")+ 'Svc, err := ' + resource.replace("_","") + '.FromRobot(machine, "my_' + resource + '_svc")\n\ncommand := map[string]interface{}{"cmd": "test", "data1": 500}\nresult, err := my' + resource.title().replace("_", "") + 'Svc.DoCommand(context.Background(), command)\n'
-                                        if resource == "slam":
-                                            go_methods[type][resource]['DoCommand']['code_sample'] = 'mySLAMService, err := slam.FromRobot(machine, "my_slam_svc")\n\ncommand := map[string]interface{}{"cmd": "test", "data1": 500}\nresult, err := mySLAMService.DoCommand(context.Background(), command)\n'
-
-                                go_methods[type][resource]['Close'] = {'proto': 'Close', \
-                                    'description': 'Close must safely shut down the resource and prevent further use. Close must be idempotent. Later reconfiguration may allow a resource to be "open" again.', \
-                                    'usage': 'Close(ctx <a href="/context">context</a>.<a href="/context#Context">Context</a>) <a href="/builtin#error">error</a>', \
-                                    'method_link': 'https://pkg.go.dev/go.viam.com/rdk/resource#Resource'}
-                                code_sample = resource_soup.find_all(lambda code_sample_tag: code_sample_tag.name == 'p' and "Close example:" in code_sample_tag.text)
-                                if code_sample:
-                                    if type == "component":
-                                        go_methods[type][resource]['Close']['code_sample'] = 'my' + resource.title().replace("_", "") + ', err := ' + go_resource_overrides.get(resource, resource) + '.FromRobot(machine, "my_' + resource + '")\n\nerr = my' + resource.title().replace("_", "") + '.Close(context.Background())\n'
-                                    else:
-                                        if resource == "base_remote_control":
-                                            go_methods[type][resource]['Close']['code_sample'] = 'baseRCService, err := baseremotecontrol.FromRobot(machine, "my_baseRCService_svc")\n\nerr := baseRCService.Close(context.Background())\n'
-                                        elif resource == "data_manager":
-                                            go_methods[type][resource]['Close']['code_sample'] = 'data, err := datamanager.FromRobot(machine, "my_data_manager")\n\nerr := data.Close(context.Background())\n'
-                                        elif resource == "navigation":
-                                            go_methods[type][resource]['Close']['code_sample'] = 'my_nav, err := navigation.FromRobot(machine, "my_nav_svc")\n\nerr := my_nav.Close(context.Background())\n'
-                                        elif resource == "mlmodel":
-                                            go_methods[type][resource]['Close']['code_sample'] = 'my_mlmodel, err := mlmodel.FromRobot(machine, "my_ml_model")\n\nerr := my_mlmodel.Close(context.Background())\n'
-                                        else:
-                                            go_methods[type][resource]['Close']['code_sample'] = 'my' + resource.title().replace("_", "") + 'Svc, err := ' + resource + '.FromRobot(machine, "my_' + resource + '_svc")\n\nerr = my' + resource.title().replace("_", "") + 'Svc.Close(context.Background())\n'
-
-                            ## Similarly, if the resource being considered inherits from resource.Actuator (Servo, for example),
-                            ## then add the two inherited methods manually: IsMoving() and Stop():
-                            if '\tresource.Actuator' in resource_interface.text:
-                                go_methods[type][resource]['IsMoving'] = {'proto': 'IsMoving', \
-                                    'description': 'IsMoving returns whether the resource is moving or not', \
-                                    'usage': 'IsMoving(ctx <a href="/context">context</a>.<a href="/context#Context">Context</a>) (<a href="/builtin#bool">bool</a>, <a href="/builtin#error">error</a>)', \
-                                    'method_link': 'https://pkg.go.dev/go.viam.com/rdk/resource#Actuator'}
-                                code_sample = resource_soup.find_all(lambda code_sample_tag: code_sample_tag.name == 'p' and "IsMoving example:" in code_sample_tag.text)
-                                if code_sample:
-                                    go_methods[type][resource]['IsMoving']['code_sample'] = code_sample[0].find_next('pre').text.replace("\t", "  ")
-                                go_methods[type][resource]['Stop'] = {'proto': 'Stop', \
-                                    'description': 'Stop stops all movement for the resource', \
-                                    'usage': 'Stop(ctx <a href="/context">context</a>.<a href="/context#Context">Context</a>, extra map[<a href="/builtin#string">string</a>]interface{}) <a href="/builtin#error">error</a>', \
-                                    'method_link': 'https://pkg.go.dev/go.viam.com/rdk/resource#Actuator'}
-                                code_sample = resource_soup.find_all(lambda code_sample_tag: code_sample_tag.name == 'p' and "Stop example:" in code_sample_tag.text)
-                                if code_sample:
-                                    go_methods[type][resource]['Stop']['code_sample'] = code_sample[0].find_next('pre').text.replace("\t", "  ")
-
-                            ## Similarly, if the resource being considered inherits from resource.Shaped (Base, for example),
-                            ## then add the one inherited method manually: Geometries():
-                            if '\tresource.Shaped' in resource_interface.text:
-                                go_methods[type][resource]['Geometries'] = {'proto': 'GetGeometries', \
-                                    'description': 'Geometries returns the list of geometries associated with the resource, in any order. The poses of the geometries reflect their current location relative to the frame of the resource.', \
-                                    'usage': 'Geometries(ctx <a href="/context">context</a>.<a href="/context#Context">Context</a>, extra map[<a href="/builtin#string">string</a>]interface{}) ([]<a href="/go.viam.com/rdk/spatialmath">spatialmath</a>.<a href="/go.viam.com/rdk/spatialmath#Geometry">Geometry</a>, <a href="/builtin#error">error</a>)', \
-                                    'method_link': 'https://pkg.go.dev/go.viam.com/rdk/resource#Shaped'}
-                                code_sample = resource_soup.find_all(lambda code_sample_tag: code_sample_tag.name == 'p' and "Geometries example:" in code_sample_tag.text)
-                                if code_sample:
-                                    go_methods[type][resource]['Geometries']['code_sample'] = code_sample[0].find_next('pre').text.replace("\t", "  ").replace('myArm', 'my{}'.format(resource.title().replace('_', ''))).replace('my_arm', 'my_{}'.format(resource)).replace('arm', resource)
-
-                            ## Similarly, if the resource being considered inherits from resource.Sensor (Movement Sensor, for example),
-                            ## then add the one inherited method manually: Readings():
-                            if '\tresource.Sensor' in resource_interface.text:
-                                go_methods[type][resource]['Readings'] = {'proto': 'GetReadings', \
-                                    'description': 'Readings return data specific to the type of sensor and can be of any type.', \
-                                    'usage': 'Readings(ctx <a href="/context">context</a>.<a href="/context#Context">Context</a>, extra map[<a href="/builtin#string">string</a>]interface{}) (map[<a href="/builtin#string">string</a>]interface{}, <a href="/builtin#error">error</a>)', \
-                                    'method_link': 'https://pkg.go.dev/go.viam.com/rdk/resource#Sensor'}
-                                code_sample = resource_soup.find_all(lambda code_sample_tag: code_sample_tag.name == 'p' and "Readings example:" in code_sample_tag.text)
-                                if code_sample:
-                                    go_methods[type][resource]['Readings']['code_sample'] = code_sample[0].find_next('pre').text.replace("\t", "  ")
-
-                ## For SLAM service only, additionally fetch data for two helper methods defined outside of the resource's interface:
-                if resource == 'slam':
-
-                    ## Fetch PointCloudMapFull:
-                    pointcloudmapfull_method_raw = soup.find_all(
-                        lambda tag: tag.name == 'div'
-                        and tag.get('class') == ['Documentation-declaration']
-                        and "PointCloudMapFull" in tag.pre.text)
-
-                    go_methods[type][resource]['PointCloudMapFull'] = {}
-                    go_methods[type][resource]['PointCloudMapFull']['proto'] = 'PointCloudMapFull'
-                    go_methods[type][resource]['PointCloudMapFull']['description'] = pointcloudmapfull_method_raw[0].pre.find_next('p').text
-                    go_methods[type][resource]['PointCloudMapFull']['usage'] = str(pointcloudmapfull_method_raw[0].pre).removeprefix('<pre>func ').removesuffix('</pre>')
-                    go_methods[type][resource]['PointCloudMapFull']['method_link'] = 'https://pkg.go.dev/go.viam.com/rdk/services/slam#PointCloudMapFull'
-
-                    ## Fetch InternalStateFull:
-                    internalstatefull_method_raw = soup.find_all(
-                        lambda tag: tag.name == 'div'
-                        and tag.get('class') == ['Documentation-declaration']
-                        and "InternalStateFull" in tag.pre.text)
-
-                    go_methods[type][resource]['InternalStateFull'] = {}
-                    go_methods[type][resource]['InternalStateFull']['proto'] = 'InternalStateFull'
-                    go_methods[type][resource]['InternalStateFull']['description'] = internalstatefull_method_raw[0].pre.find_next('p').text
-                    go_methods[type][resource]['InternalStateFull']['usage'] = str(internalstatefull_method_raw[0].pre).removeprefix('<pre>func ').removesuffix('</pre>')
-                    go_methods[type][resource]['InternalStateFull']['method_link'] = 'https://pkg.go.dev/go.viam.com/rdk/services/slam#InternalStateFull'
-
-                ## We have finished looping through all scraped Go methods. Write the go_methods dictionary
-                ## in its entirety to the all_methods dictionary using "go" as the key:
-                all_methods["go"] = go_methods
-
-            ## Assemble workaround data object for DoCommand for Go generic component and service.
-            ## Using code sample and method_link from resource.Resource, because these cannot be found
-            ## in Go docs for these resources:
-            elif sdk == "go" and (resource == "generic_component" or resource == "generic_service"):
-
-                go_methods[type][resource]['DoCommand'] = {}
-                go_methods[type][resource]['DoCommand'] = {'proto': 'DoCommand', \
-                                    'description': 'DoCommand sends/receives arbitrary data.', \
-                                    'usage': 'DoCommand(ctx <a href="/context">context</a>.<a href="/context#Context">Context</a>, cmd map[<a href="/builtin#string">string</a>]interface{}) (map[<a href="/builtin#string">string</a>]interface{}, <a href="/builtin#error">error</a>)', \
-                                    'method_link': 'https://pkg.go.dev/go.viam.com/rdk/resource#Resource', \
-                                    'code_sample': 'my' + resource.title().replace("_", "") + ', err := generic.FromRobot(machine, "my_' + resource.lower() + '")\n\ncommand := map[string]interface{}{"cmd": "test", "data1": 500}\nresult, err := my' + resource.title().replace("_", "") + '.DoCommand(context.Background(), command)\n'}
-                if resource == "generic_service":
-                    go_methods[type][resource]['DoCommand']['code_sample'] = 'myGenericService, err := generic.FromRobot(machine, "my_generic_service")\n\ncommand := map[string]interface{}{"cmd": "test", "data1": 500}\nresult, err := myGenericService.DoCommand(context.Background(), command)\n'
-
-                all_methods["go"] = go_methods
-
-            elif sdk == "go" and type == "app":
-               ##Go SDK has no APP API!
-               pass
-
-            ## Scrape each parent method tag and all contained child tags for Python by resource.
-            ## TEMP: Manually exclude Base Remote Control Service (Go only) and Data Manager Service (Go + Flutter only).
-            ## TODO: Handle resources with 0 implemented methods for this SDK better.
-            elif sdk == "python" and resource != 'base_remote_control' and resource != 'data_manager':
-                soup = make_soup(url)
-                python_methods_raw = soup.find_all("dl", class_="py method")
-
-                ## Loop through scraped tags and select salient data:
-                for tag in python_methods_raw:
-
-                    ## Create new empty dictionary for this specific method, to be appended to ongoing python_methods dictionary,
-                    ## in form: python_methods[type][resource][method_name] = this_method_dict
-                    this_method_dict = {}
-
-                    id = tag.find("dt", class_="sig sig-object py").get("id")
-
-                    if not id.endswith(".get_operation") \
-                    and not id.endswith(".from_proto") and not id.endswith(".to_proto") \
-                    and not id.endswith(".from_string") and not id.endswith("__") \
-                    and not id.endswith("HasField") and not id.endswith("WhichOneof") \
-                    and not id in python_ignore_apis:
-
-                        ## Determine method name, but don't save to dictionary as value; we use it as a key instead:
-                        method_name = id.rsplit('.',1)[1]
-
-                        ## Look up method_name in proto_map file, and return matching proto:
-                        with open(proto_map_file, 'r') as f:
-                            for row in f:
-                                if not row.startswith('#') \
-                                and row.startswith(resource + ',') \
-                                and row.split(',')[3] == method_name:
-                                    this_method_dict["proto"] = row.split(',')[1]
-
-                        ## Determine method description, stripping newlines. If not present, skip:
-                        if tag.find('dd').p:
-                            this_method_dict["description"] = tag.find('dd').p.text.replace("\n", " ")
-
-                        ## Determine method direct link, no need to parse for it, it's inferrable.
-                        ## If we are scraping from a local staging instance, replace host and port with upstream link target URL:
-                        if python_staging_url != '':
-                            this_method_dict["method_link"] = str(url + "#" + id).replace(python_staging_url, 'https://python.viam.dev')
-                        else:
-                            this_method_dict["method_link"] = url + "#" + id
-
-                        ## Assemble array of all tags which contain parameters for this method:
-                        ## METHODOLOGY: tag: em, class: sig-param, and not * or **kwargs:
-                        parameter_tags = tag.find_all(
-                            lambda tag: tag.name == 'em'
-                            and tag.get('class') == ['sig-param']
-                            and not regex.search(r'\*', tag.text))
-
-                        if len(parameter_tags) != 0:
-
-                            ## Create new empty dictionary for this_method_dict named "parameters":
-                            this_method_dict["parameters"] = {}
-
-                            # Iterate through each parameter found for this method:
-                            for parameter_tag in parameter_tags:
-                                ## Create new empty dictionary this_method_parameters_dict to house all parameter
-                                ## keys for this method, to allow for multiple parameters. Also resets the
-                                ## previous parameter's data when looping through multiple parameters:
-                                this_method_parameters_dict = {}
-
-                                ## Determine parameter name, but don't save to dictionary as value; we use it as a key instead:
-                                param_name = parameter_tag.find('span', class_="pre").text
-
-                                ## Determine parameter type:
-                                param_type = parameter_tag.find_all('span', class_='n')[1].text
-
-                                param_default = parameter_tag.select_one('span.default_value')
-                                if param_default:
-                                    param_default = param_default.text
-                                else:
-                                    param_default = None
-
-                                if param_default == "''":  # Check for empty string default
-                                    this_method_parameters_dict["optional"] = True
-                                ## Determine if this parameter is optional, and strip off ' | None' syntax if so:
-                                elif param_type.endswith(' | None'):
-                                    this_method_parameters_dict["optional"] = True
-                                    param_type = param_type.replace(' | None', "")
-                                else:
-                                    this_method_parameters_dict["optional"] = False
-
-                                ## 'Extra' params do not appear in "Parameters" section (Except for PySDK > Motion Service),
-                                ## so we must populate this param's content manually:
-                                if param_name == 'extra':
-
-                                    this_method_parameters_dict["param_description"] = "Extra options to pass to the underlying RPC call."
-                                    this_method_parameters_dict["param_usage"] = "extra (Mapping[str, Any]) - Extra options to pass to the underlying RPC call."
-                                    this_method_parameters_dict["param_type"] = "Mapping[[str](https://docs.python.org/3/library/stdtypes.html#text-sequence-type-str), Any]"
-
-                                ## 'Timeout' params do not appear in "Parameters" section (Except for PySDK > Motion Service),
-                                ## so we must populate this param's content manually:
-                                elif param_name == 'timeout':
-
-                                    this_method_parameters_dict["param_description"] = "An option to set how long to wait (in seconds) before calling a time-out and closing the underlying RPC call."
-                                    this_method_parameters_dict["param_usage"] = "timeout (float) - An option to set how long to wait (in seconds) before calling a time-out and closing the underlying RPC call."
-                                    this_method_parameters_dict["param_type"] = "[float](https://docs.python.org/3/library/stdtypes.html#numeric-types-int-float-complex)"
-
-                                ## Initial method usage syntax and Parameters section do not agree on param_type for do_command.
-                                ## Manually override with correct values, for param 'command' only:
-                                elif method_name == 'do_command' and param_name == 'command':
-
-                                    this_method_parameters_dict["param_description"] = "The command to execute"
-                                    this_method_parameters_dict["param_usage"] = "command (Mapping[str, ValueTypes]) – The command to execute"
-                                    this_method_parameters_dict["param_type"] = "Mapping[[str](https://docs.python.org/3/library/stdtypes.html#text-sequence-type-str), ValueTypes]"
-
-                                else:
-
-                                    ## First, check python_datatype_links array for manually-mapped data type links.
-                                    ## These will override all other ways of determining data type links:
-                                    linked_param_type = ""
-                                    linked_param_type = link_data_types('python', param_type)
-
-                                    ## If link_data_types() returns a linked data type string, use that.
-                                    ## Otherwise, check the scraped parameter_tag for the link:
-                                    if linked_param_type != "":
-                                        this_method_parameters_dict["param_type"] = linked_param_type
-                                    elif parameter_tag.find('a', class_="reference internal"):
-                                        param_type_link_raw = parameter_tag.find('a', class_="reference internal").get("href")
-
-                                        ## Parameter type link is an anchor link:
-                                        if param_type_link_raw.startswith('#'):
-                                            this_method_parameters_dict["param_type"] = '[' + param_type + '](' + url + param_type_link_raw + ')'
-                                        ## Parameter type link is a relative link, in one of three forms:
-                                        elif param_type_link_raw.startswith('../../../'):
-                                            this_method_parameters_dict["param_type"] = '[' + param_type + '](' + sdk_url + "/autoapi/viam/" + param_type_link_raw.replace('../../../', '')+ ')'
-                                        elif param_type_link_raw.startswith('../../'):
-                                            this_method_parameters_dict["param_type"] = '[' + param_type + '](' + sdk_url + "/autoapi/viam/" + param_type_link_raw.replace('../../', '')+ ')'
-                                        elif param_type_link_raw.startswith('../'):
-                                            this_method_parameters_dict["param_type"] = '[' + param_type + '](' + sdk_url + "/autoapi/viam/" + param_type_link_raw.replace('../', '')+ ')'
-
-                                    else:
-                                        this_method_parameters_dict["param_type"] = param_type
-
-                                    ## Get parameter usage and description, if method contains a "Parameters" section. Otherwise omit.
-                                    ## NOTE: We can't just use the initial param content as found above, because it does not contain descriptions,
-                                    ## and we can't just use this "Parameters" section, because it does not (usually) contain things like `extra` and `timeout`.
-                                    ## METHODOLOGY: Find parent <p> tag around matching <strong>param_name</strong> tag which contains this data.
-                                    ##   Determining by <strong> tags allows matching parameters regardless whether they are
-                                    ##   presented in <p> tags (single param) or <li> tags (multiple params):
-                                    for strong_tag in tag.find_all('strong'):
-                                        ## We have to explicitly exclude extra and timeout from this loop also,
-                                        ## because Python Motion service includes them explicitly as well:
-                                        if param_name != 'extra' and \
-                                            param_name != 'timeout' and \
-                                            strong_tag.text == param_name:
-
-                                            ## OPTION: Get just the parameter description, stripping all newlines:
-                                            this_method_parameters_dict["param_description"] = regex.split(r" – ", strong_tag.parent.text)[1].replace("\n", " ")
-
-                                            ## OPTION: Get full parameter usage string, stripping all newlines:
-                                            ## NOTE: Currently unused.
-                                            this_method_parameters_dict['param_usage'] = strong_tag.parent.text.replace("\n", " ")
-
-                                            ## Some params provide data type links in Parameters section only, not initial usage.
-                                            ## Get that here if so:
-                                            if strong_tag.parent.find('a', class_="reference internal"):
-                                                param_type_link_raw = strong_tag.parent.find('a', class_="reference internal").get("href")
-                                                ## Parameter type link is an anchor link:
-                                                if param_type_link_raw.startswith('#') and python_staging_url == '':
-                                                    this_method_parameters_dict["param_type"] = '[' + param_type + '](' + url + param_type_link_raw + ')'
-                                                elif param_type_link_raw.startswith('#') and python_staging_url != '':
-                                                    this_method_parameters_dict["param_type"] = '[' + param_type + '](' + url.replace(python_staging_url, sdk_url) + param_type_link_raw + ')'
-                                                ## Parameter type link is a relative link, beginning with 1 - 3 instances of '../'
-                                                ## Convert to an absolute link:
-                                                elif param_type_link_raw.startswith('../') \
-                                                    and not param_type_link_raw.startswith('../../'):
-                                                    if resource in python_resource_overrides:
-                                                        linkable_resource = python_resource_overrides[resource]
-                                                    else:
-                                                        linkable_resource = resource
-                                                    this_method_parameters_dict["param_type"] = f'[{param_type}]({sdk_url}/autoapi/viam/{type}s/{linkable_resource}/' + param_type_link_raw.replace('../', '')+ ')'
-                                                elif param_type_link_raw.startswith('../'):
-                                                    this_method_parameters_dict["param_type"] = '[' + param_type + '](' + sdk_url + "/autoapi/viam/" + param_type_link_raw.replace('../', '')+ ')'
-
-                                            ## Unable to determine parameter description, neither timeout or extra, nor matching to any
-                                            ## param in initial method usage string. Usually this means a non-param (like error raised),
-                                            ## but if we are missing expected param descriptions, expand this section to catch them.
-                                            else:
-                                                ## No-op:
-                                                pass
-
-                                this_method_dict["parameters"][param_name] = this_method_parameters_dict
-
-                        ## Get single tag containing the return for this method:
-                        return_tag = tag.find('span', class_='sig-return')
-
-                        ## Parse return for this method:
-                        ## METHODOLOGY: Some methods explicitly state that they return "None", others just omit the field.
-                        ##   Either way, ensure we only write a return to this_method_dict if an actual return is present:
-                        if return_tag and return_tag.find('span', class_='pre').text != "None":
-
-                            ## Create new empty dictionary for this_method_dict named "return":
-                            this_method_dict["return"] = {}
-
-                            ## OPTION: Get return_type by explicit key name:
-                            return_type = return_tag.find('span', class_="sig-return-typehint").text
-                            this_method_dict["return"]["return_type"] = return_type
-
-                            linked_return_type = ""
-                            linked_return_type = link_data_types('python', return_type)
-
-                            if linked_return_type != "":
-                                this_method_dict["return"]["return_type"] = linked_return_type
-                            elif return_tag.find('a', class_="reference internal"):
-
-                                ## TODO: Only grabbing the first link encountered, but a few methods return a tuple of two linked data types.
-                                ## Handling those via link_data_types() with manual entries in python_datatype_links for now,
-                                ## But there's room for a more elegant solution:
-                                return_type_link_raw = return_tag.find('a', class_="reference internal").get("href")
-
-                                ## Return type link is an anchor link:
-                                if return_type_link_raw.startswith('#') and python_staging_url == '':
-                                    this_method_dict["return"]["return_type"] = '[' + return_type + '](' + url + return_type_link_raw + ')'
-                                elif return_type_link_raw.startswith('#') and python_staging_url != '':
-                                    this_method_dict["return"]["return_type"] = '[' + return_type + '](' + url.replace(python_staging_url, sdk_url) + return_type_link_raw + ')'
-                                ## Return type link is a relative link, beginning with 1 - 3 instances of '../'
-                                ## Convert to an absolute link:
-                                elif return_type_link_raw.startswith('../') \
-                                    and not return_type_link_raw.startswith('../../'):
-                                    if resource in python_resource_overrides:
-                                        linkable_resource = python_resource_overrides[resource]
-                                    else:
-                                        linkable_resource = resource
-                                    this_method_dict["return"]["return_type"] = f'[{return_type}]({sdk_url}/autoapi/viam/{type}s/{linkable_resource}/' + return_type_link_raw.replace('../', '')+ ')'
-                                elif return_type_link_raw.startswith('../'):
-                                    this_method_dict["return"]["return_type"] = '[' + return_type + '](' + sdk_url + "/autoapi/viam/" + return_type_link_raw.replace('../', '')+ ')'
-
-                            ## OPTION: Get full return usage, including type info and html links if present, stripping all newlines:
-                            ## NOTE: Currently unused.
-                            this_method_dict["return"]["return_usage"] = str(return_tag.find('span', class_="sig-return-typehint")).replace("\n", " ")
-
-                            ## Get return description from "Returns" section if present:
-                            if tag.find(string="Returns"):
-
-                                ## METHODOLOGY: Return description is always in <dd> tag with class either 'field-odd' or 'field-even',
-                                ##   but this is the same as "Parameters" description. Returns differ by not being enclosed in <strong> tags:
-                                return_description_raw = tag.find_all(
-                                       lambda tag: tag.name == 'dd'
-                                       and not tag.find_next('p').strong
-                                       and (tag.get('class') == ['field-odd']
-                                       or tag.get('class') == ['field-even']))
-
-                                ## Append to ongoing this_method_dict, stripping newlines:
-                                this_method_dict["return"]["return_description"] = return_description_raw[0].p.text.replace("\n", " ")
-
-                        ## If method has a "Raises" section, determine method errors raised:
-                        if tag.find(string="Raises"):
-
-                            ## Create new empty dictionary for this_method_dict named "raises",
-                            ## and new empty dictionary this_method_raises_dict to house all errors raised
-                            ## keys for this method, to allow for multiple errors raised:
-                            this_method_dict["raises"] = {}
-
-                            ## Iterate through all <strong> tags in method tag:
-                            for strong_tag in tag.find_all('strong'):
-
-                                ## Create new empty dictionary this_method_raises_dict to house all raises (errors)
-                                ## keys for this method, to allow for multiple errors raised. Also resets the
-                                ## previous error's data when looping through multiple errors:
-                                this_method_raises_dict = {}
-
-                                ## Determine if this <strong> tag is preceded by a <dt> tag containing the text "Raises". Otherwise omit.
-                                ## METHODOLOGY: Find previous <dt> tag before matching <strong>param_name</strong> tag which contains this data.
-                                ##   Determining by <strong> tags allows matching parameters regardless whether they are
-                                ##   presented in <p> tags (single error raised) or <li> tags (multiple errors raised):
-                                if strong_tag.find_previous('dt').text == "Raises:":
-
-                                    ## Split contained text at " - " to get first half, which is just the error name:
-                                    raises_name = regex.split(r" – ", strong_tag.parent.text)[0]
-
-                                    ## Process remaining data fields depending on whether the error data type is linked or not.
-                                    ## If the error includes a linked data type:
-                                    if strong_tag.parent.name == 'a':
-
-                                        ## OPTION: Get full error raised usage, including type info and html links if present.
-                                        ## NOTE: Currently unused.
-                                        raises_usage = str(strong_tag.parent.parent).replace("\n", " ")
-                                        raises_link = strong_tag.parent.get('href')
-
-                                        ## Replace the scraped relative link with a full URL, in one of two forms:
-                                        ## Scraped link is an anchor link:
-                                        if raises_link.startswith('#'):
-                                            this_method_raises_dict["raises_usage"] = regex.sub(r'href=".*"', 'href="' + url + raises_link + '"', raises_usage)
-                                        ## Scraped link is a relative link:
-                                        elif raises_link.startswith('../'):
-                                            this_method_raises_dict["raises_usage"] = regex.sub(r'href=".*"', 'href="' + sdk_url + '/autoapi/viam/' + raises_link.replace('../', '') + '"', raises_usage)
-
-                                        ## OPTION: Determine error raised description, stripping any newlines:
-                                        this_method_raises_dict["raises_description"] = regex.split(r" – ", strong_tag.parent.parent.text)[1].replace("\n", " ")
-
-                                    ## If the error does not include a linked data type:
-                                    else:
-                                        ## OPTION: Get full error raised usage:
-                                        ## NOTE: Currently unused.
-                                        this_method_raises_dict["raises_usage"] = str(strong_tag.parent).replace("\n", " ")
-
-                                        ## OPTION: Determine error raised description, stripping any newlines:
-                                        this_method_raises_dict["raises_description"] = regex.split(r" – ", strong_tag.parent.text)[1].replace("\n", " ")
-
-                                    ## Add all values for this raised error to this_method_dict by raises_name:
-                                    this_method_dict["raises"][raises_name] = this_method_raises_dict
-
-                        ## Determine if a code sample is provided for this method:
-                        if tag.find('div', class_="highlight"):
-
-                            ## Fetch code sample raw text, preserving newlines but stripping all formatting.
-                            ## This string should be suitable for feeding into any python formatter to get proper form:
-                            this_method_dict["code_sample"] = tag.find('div', class_="highlight").pre.text
-
-                        # For do_command, update the code sample to be more specific:
-                        if method_name == "do_command":
-                            if type == "service":
-                                this_method_dict["code_sample"] = this_method_dict["code_sample"].replace("  # replace SERVICE with the appropriate class", "")
-                                service_class = resource.title() + "Client"
-                                if resource == "mlmodel":
-                                    service_class = "MLModelClient"
-                                elif resource == "generic_service":
-                                    service_class = "Generic"
-                                elif resource == "slam":
-                                    service_class = "SLAMClient"
-                                this_method_dict["code_sample"] = this_method_dict["code_sample"].replace("SERVICE", service_class)
-                                this_method_dict["code_sample"] = this_method_dict["code_sample"].replace("# Can be used with any resource, using the motion service as an example\n", "")
-                                if not resource == "generic_service":
-                                    this_method_dict["code_sample"] = this_method_dict["code_sample"].replace("service", "my_{}_svc".format(resource))
-                                if not resource in ["motion"]:
-                                    this_method_dict["code_sample"] = this_method_dict["code_sample"].replace("SERVICE", service_class).replace("builtin", "my_{}_svc".format(resource))
-                            else:
-                                if resource == "generic_component":
-                                    this_method_dict["code_sample"] = 'my_generic_component = Generic.from_robot(robot=machine, name="my_generic_component")\n' + this_method_dict["code_sample"].replace("component", "my_generic_component")
-                                else:
-                                    this_method_dict["code_sample"] = 'my_{} = {}.from_robot(robot=machine, name="my_{}")\n'.format(resource, resource.title().replace("_", ""), resource) + this_method_dict["code_sample"].replace("component", "my_{}".format(resource))
-
-                        # For get_geometries, update the code sample to be more specific:
-                        if method_name == "get_geometries":
-                            if type == "service":
-                                if resource == "generic_service":
-                                    this_method_dict["code_sample"] = 'my_{} = {}.from_robot(robot=machine, name="my_{}")\n'.format(resource, resource.title(), resource) + this_method_dict["code_sample"].replace("component", "my_{}".format(resource))
-                                else:
-                                    this_method_dict["code_sample"] = 'my_{} = {}.from_robot(robot=machine, name="my_{}_svc")\n'.format(resource, resource.title(), resource) + this_method_dict["code_sample"].replace("component", "my_{}_svc".format(resource))
-                            else:
-                                if resource == "generic_component":
-                                    this_method_dict["code_sample"] = 'my_generic_component = Generic.from_robot(robot=machine, name="my_generic_component")\n' + this_method_dict["code_sample"].replace("component", "my_generic_component")
-                                elif resource == "input_controller":
-                                    this_method_dict["code_sample"] = 'my_controller = Controller.from_robot(robot=machine, name="my_controller")\n' + this_method_dict["code_sample"].replace("component", "my_controller")
-                                else:
-                                    this_method_dict["code_sample"] = 'my_{} = {}.from_robot(robot=machine, name="my_{}")\n'.format(resource, resource.title().replace("_", ""), resource) + this_method_dict["code_sample"].replace("component", "my_{}".format(resource))
-
-                        # For close, update the code sample to be more specific:
-                        if method_name == "close":
-                            if type == "service":
-                                if resource == "generic_service":
-                                    this_method_dict["code_sample"] = 'my_{} = Generic.from_robot(robot=machine, name="my_{}")\nawait my_{}.close()\n'.format(resource, resource, resource)
-                                elif resource == "mlmodel":
-                                    this_method_dict["code_sample"] = 'my_{}_svc = MLModelClient.from_robot(robot=machine, name="my_{}_svc")\nawait my_{}_svc.close()\n'.format(resource, resource, resource)
-                                elif resource == "slam":
-                                    this_method_dict["code_sample"] = 'my_{}_svc = SLAMClient.from_robot(robot=machine, name="my_{}_svc")\nawait my_{}_svc.close()\n'.format(resource, resource, resource)
-                                else:
-                                    this_method_dict["code_sample"] = 'my_{}_svc = {}Client.from_robot(robot=machine, name="my_{}_svc")\nawait my_{}_svc.close()\n'.format(resource, resource.title(), resource, resource)
-                            elif type == "component":
-                                if resource == "generic_component":
-                                    this_method_dict["code_sample"] = 'my_generic_component = Generic.from_robot(robot=machine, name="my_generic_component")\nawait my_generic_component.close()\n'
-                                elif resource == "input_controller":
-                                    this_method_dict["code_sample"] = 'my_controller = Controller.from_robot(robot=machine, name="my_controller")\nawait my_controller.close()\n'
-                                else:
-                                    this_method_dict["code_sample"] = 'my_{} = {}.from_robot(robot=machine, name="my_{}")\nawait my_{}.close()\n'.format(resource, resource.title().replace("_", ""), resource, resource)
-
-                        # For get_resource_name, update the code sample to be more specific:
-                        if method_name == "get_resource_name":
-                            if type == "service":
-                                service_class = resource.title().replace('_','') + "Client"
-                                if resource == "mlmodel":
-                                    service_class = "MLModelClient"
-                                elif resource == "generic_service":
-                                    service_class = "Generic"
-                                elif resource == "slam":
-                                    service_class = "SLAMClient"
-                                if resource == "generic_service":
-                                    this_method_dict["code_sample"] = 'my_{}_name = {}.get_resource_name("my_{}")\n'.format(resource, service_class ,resource)
-                                else:
-                                    this_method_dict["code_sample"] = 'my_{}_svc_name = {}.get_resource_name("my_{}_svc")\n'.format(resource, service_class ,resource)
-                            else:
-                                if resource == "input_controller":
-                                    this_method_dict["code_sample"] = 'my_input_controller_name = Controller.get_resource_name("my_input_controller")\n'
-                                elif resource == "generic_component":
-                                    this_method_dict["code_sample"] = 'my_generic_component_name = Generic.get_resource_name("my_generic_component")\n'
-                                else:
-                                    this_method_dict["code_sample"] = 'my_{}_name = {}.get_resource_name("my_{}")\n'.format(resource, resource.title().replace('_','') ,resource)
-
-                        ## We have finished collecting all data for this method. Write the this_method_dict dictionary
-                        ## in its entirety to the python_methods dictionary by type (like 'component'), by resource (like 'arm'),
-                        ## using the method_name as key:
-
-                        python_methods[type][resource][method_name] = this_method_dict
-
-                ## We have finished looping through all scraped Python methods. Write the python_methods dictionary
-                ## in its entirety to the all_methods dictionary using "python" as the key:
-                all_methods["python"] = python_methods
-
-            ## Scrape each parent method tag and all contained child tags for Flutter by resource.
-            ## TEMP: Manually exclude Base Remote Control Service (Go only).
-            ## TODO: Handle resources with 0 implemented methods for this SDK better.
-            elif sdk == "flutter" and resource != 'base_remote_control' and resource != 'button' and resource != 'encoder' \
-                and resource != 'input_controller' and resource != 'switch' and resource != 'data_manager' \
-                and resource != 'generic_service' and resource != 'mlmodel' and resource != 'motion' \
-                and resource != 'navigation' and resource != 'slam':
-                soup = make_soup(url)
-
-                if resource in flutter_resource_overrides:
-                    flutter_resource = flutter_resource_overrides[resource]
-                else:
-                    flutter_resource = resource.capitalize()
-                ## Limit matched class to either 'callable' or 'callable inherited' and remove the constructor (proto id) itself:
-                flutter_methods_raw = soup.find_all(
-                    lambda tag: tag.name == 'dt'
-                    and not tag.get('id') == flutter_resource
-                    and tag.has_attr("class")
-                    and "callable" in tag.get("class"))
-
-                ## Loop through scraped tags and select salient data:
-                for tag in flutter_methods_raw:
-
-                    ## Create new empty dictionary for this specific method, to be appended to ongoing flutter_methods dictionary,
-                    ## in form: flutter_methods[type][resource][method_name] = this_method_dict
-                    this_method_dict = {}
-
-                    method_name = tag.get('id')
-
-                    if not method_name in flutter_ignore_apis:
-
-                        ## Look up method_name in proto_map file, and return matching proto:
-                        with open(proto_map_file, 'r') as f:
-                            for row in f:
-                                ## Because Flutter is the final entry in the mapping CSV, we must also rstrip() to
-                                ## strip the trailing newline (\n) off the row itself:
-                                row = row.rstrip()
-
-                                if not row.startswith('#') \
-                                and row.startswith(resource + ',') \
-                                and row.split(',')[5] == method_name:
-                                    this_method_dict["proto"] = row.split(',')[1]
-
-                        ## Determine method link:
-                        method_link = tag.find("span", class_="name").a['href'].replace("..", sdk_url)
-                        this_method_dict["method_link"] = method_link
-
-                        ## While some method info is available to us on this current Flutter SDK page, the code sample is only found on the
-                        ## method_link page. So we scrape that page for everything:
-                        method_soup = make_soup(method_link)
-
-                        ## Method description and code samples are both found within the same section tag:
-                        desc_or_code_sample = method_soup.find('section', class_ = 'desc markdown')
-
-                        if desc_or_code_sample:
-                            if desc_or_code_sample.p:
-                                this_method_dict["method_description"] = desc_or_code_sample.p.text
-                            if desc_or_code_sample.pre:
-                                this_method_dict["code_sample"] = desc_or_code_sample.pre.text
-
-                        parameter_tags = method_soup.find_all(
-                            lambda tag: tag.name == 'span'
-                            and tag.get('class') == ['parameter'])
-
-                        ## Parse parameters, if any are found:
-                        if len(parameter_tags) != 0:
-
-                            ## Create new empty dictionary for this_method_dict named "parameters":
-                            this_method_dict["parameters"] = {}
-
-                            optional = False
-
-                            # If there is a curly brace before the parameter list all parameters are optional
-                            prev = parameter_tags[0].find_previous('section')
-                            if prev:
-                                if "{<ol" in str(prev):
-                                    optional = True
-
-                            for parameter_tag in parameter_tags:
-
-                                ## Create new empty dictionary this_method_parameters_dict to house all parameter
-                                ## keys for this method, to allow for multiple parameters. Also resets the
-                                ## previous parameter's data when looping through multiple parameters:
-                                this_method_parameters_dict = {}
-
-                                ## Parse for param name and usage string, convert to string (for markdownify):
-                                param_name = parameter_tag.find('span', class_ = 'parameter-name').text
-                                param_usage = str(parameter_tag.find('span', class_ = 'type-annotation')).replace('>>', '>\\>')
-
-                                ## Markdownify parameter usage and replace relative links with absolute:
-                                formatted_param_usage = md(param_usage, strip=['wbr']).replace("../../", "https://flutter.viam.dev/")
-                                this_method_parameters_dict["param_usage"] = formatted_param_usage
-
-                                # if a parameter is optional the previous parameter has a curly brace before it
-                                this_method_parameters_dict["optional"] = optional
-
-                                this_method_dict["parameters"][param_name] = this_method_parameters_dict
-
-                                if parameter_tag.text.endswith("{"):
-                                    optional = True
-
-                        return_tags = method_soup.find_all(
-                            lambda tag: tag.name == 'span'
-                            and tag.get('class') == ['returntype'])
-
-                        if len(return_tags) != 0:
-
-                            ## Create new empty dictionary for this_method_dict named "return":
-                            this_method_dict["return"] = {}
-
-                            for return_tag in return_tags:
-
-                                ## Create new empty dictionary this_method_returns_dict to house all return
-                                ## keys for this method, to allow for multiple returns. Also resets the
-                                ## previous return's data when looping through multiple returns:
-                                this_method_return_dict = {}
-
-                                # Parse return usage string, convert to string (for markdownify):
-                                return_usage = str(return_tag)
-
-                                ## Markdownify return usage and replace relative links with absolute:
-                                formatted_return_usage = md(return_usage, strip=['wbr']).replace("../../", "https://flutter.viam.dev/")
-                                this_method_return_dict["return_usage"] = formatted_return_usage.replace('>>', '>\\>')
-
-                                # Parse return type:
-                                if return_tag.find('span', class_ = 'type-parameter'):
-                                    return_type = return_tag.find('span', class_ = 'type-parameter').text
-                                else:
-                                    return_type = return_tag.text
-
-                                this_method_dict["return"][return_type] = this_method_return_dict
-
-                        flutter_methods[type][resource][method_name] = this_method_dict
-
-                ## We have finished looping through all scraped Flutter methods. Write the flutter_methods dictionary
-                ## in its entirety to the all_methods dictionary using "flutter" as the key:
-                all_methods["flutter"] = flutter_methods
-
-            elif sdk == "flutter" and type == "app":
-                ##Flutter SDK has no APP API!
-                pass
-
-            else:
-                ## Good code would never get here.
-                ## This code gets here when facing a resource with 0 implemented methods for
-                ## the SDK we're looping for.
-                ## TODO: Fix so resources with 0 implemented methods for an SDK are silently
-                ## skipped without requiring manual exclusion in parse().
-                pass
 
     return all_methods
